UPD. update copyProperties function

+++ backend/src/main/docker/mysql/dump.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `users` (
-  `id` int auto_increment,
+  `id` BIGINT auto_increment,
   `name` varchar(255) NOT NULL,
   PRIMARY KEY (`id`)
 );
